+++ packages/shared/extension-preload/src/types/api.d.ts
@@ -975,6 +975,8 @@ declare global {
        * (for example, macOS).
        */
       title: string
+      /** Allow to select folder, defaults to `false`. */
+      canSelectFolder?: boolean
     }
 
     /** 环境相关 */