+++ packages/web-server/publish/changeLog.md
@@ -7,6 +7,7 @@
 
 ### Fixed
 
+- Fixed an issue where the file save dialog API could not select folders ([#285](https://github.com/any-listen/any-listen/issues/285)).
 - Fixed an issue where some lyrics did not wrap properly.
 
 ---
@@ -20,4 +21,5 @@
 
 ### 修复
 
+- 修复文件保存弹窗 API 无法选择文件夹的问题（[#285](https://github.com/any-listen/any-listen/issues/285)）。
 - 修复某些歌词不换行的问题。